@@ -192,6 +192,7 @@ _c_headers = """#include <math.h>
 /*Some functions supported by numpy that aren't included in math.h
  * names and arguments match the wrappers used in rxdmath.py
  */
+extern "C" {
 double factorial(const double);
 double degrees(const double);
 void radians(const double, double*);
@@ -519,35 +520,35 @@ def _find_librxdmath():
     return dll
 
 
-def _c_compile(formula):
+def _cxx_compile(formula):
     filename = "rxddll" + str(uuid.uuid1())
-    with open(filename + ".c", "w") as f:
+    with open(filename + ".cpp", "w") as f:
         f.write(formula)
     math_library = "-lm"
     fpic = "-fPIC"
     try:
-        gcc = os.environ["CC"]
+        gcc = os.environ["CXX"]
     except:
         # when running on windows try and used the gcc included with NEURON
         if sys.platform.lower().startswith("win"):
             math_library = ""
             fpic = ""
             gcc = os.path.join(
-                h.neuronhome(), "mingw", "mingw64", "bin", "x86_64-w64-mingw32-gcc.exe"
+                h.neuronhome(), "mingw", "mingw64", "bin", "x86_64-w64-mingw32-g++.exe"
             )
             if not os.path.isfile(gcc):
                 raise RxDException(
-                    "unable to locate a C compiler. Please `set CC=<path to C compiler>`"
+                    "unable to locate a CXX compiler. Please `set CXX=<path to CXX compiler>`"
                 )
         else:
-            gcc = "gcc"
+            gcc = "g++"
     # TODO: Check this works on non-Linux machines
     gcc_cmd = "%s -I%s -I%s " % (
         gcc,
         sysconfig.get_path("include"),
         os.path.join(h.neuronhome(), "..", "..", "include", "nrn"),
     )
-    gcc_cmd += "-shared %s  %s.c %s " % (fpic, filename, _find_librxdmath())
+    gcc_cmd += "-shared %s  %s.cpp %s " % (fpic, filename, _find_librxdmath())
     gcc_cmd += "-o %s.so %s" % (filename, math_library)
     if sys.platform.lower().startswith("win"):
         my_path = os.getenv("PATH")
@@ -568,7 +569,7 @@ def _c_compile(formula):
         ctypes.POINTER(ctypes.c_double),
     ]
     reaction.restype = ctypes.c_double
-    os.remove(filename + ".c")
+    os.remove(filename + ".cpp")
     if sys.platform.lower().startswith("win"):
         # cannot remove dll that are in use
         _windows_dll.append(weakref.ref(dll))
@@ -1467,7 +1468,7 @@ def _compile_reactions():
                                 operator,
                                 summed_mults[idx],
                             )
-            fxn_string += "\n}\n"
+            fxn_string += "\n}\n}\n"
             register_rate(
                 creg.num_species,
                 creg.num_params,
@@ -1481,7 +1482,7 @@ def _compile_reactions():
                 mc_mult_count,
                 numpy.array(mc_mult_list, dtype=ctypes.c_double),
                 _list_to_pyobject_array(creg._vptrs),
-                _c_compile(fxn_string),
+                _cxx_compile(fxn_string),
             )
 
     # Setup intracellular 3D reactions
@@ -1710,7 +1711,7 @@ def _compile_reactions():
                             r._mult[idx],
                         )
                         idx += 1
-            fxn_string += "\n}\n"
+            fxn_string += "\n}\n}\n"
             for i, ele in enumerate(mults):
                 if ele == []:
                     mults[i] = numpy.ones(len(reg._xs))
@@ -1723,7 +1724,7 @@ def _compile_reactions():
                 numpy.asarray(mc3d_indices_start, dtype=numpy.uint64),
                 mc3d_region_size,
                 numpy.asarray(mults),
-                _c_compile(fxn_string),
+                _cxx_compile(fxn_string),
             )
     # Setup extracellular reactions
     if len(ecs_regions_inv) > 0:
@@ -1833,13 +1834,13 @@ def _compile_reactions():
                             r._mult[idx],
                         )
                         idx += 1
-            fxn_string += "\n}\n"
+            fxn_string += "\n}\n}\n"
             ecs_register_reaction(
                 0,
                 len(all_gids),
                 len(param_gids),
                 _list_to_cint_array(all_gids + param_gids),
-                _c_compile(fxn_string),
+                _cxx_compile(fxn_string),
             )
 
 
