@@ -48,6 +48,7 @@ extern Image* gif_image(const char*);
 
 #include "classreg.h"
 #include "gui-redirect.h"
+#include "treeset.h"
 
 extern Object** (*nrnpy_gui_helper_)(const char* name, Object* obj);
 extern double (*nrnpy_object_to_double_)(Object*);
@@ -1497,7 +1498,6 @@ Graph::Graph(bool b)
 
 extern "C" {
 extern void hoc_free_list(Symlist**);
-extern double* nrn_recalc_ptr(double*);
 }  // extern "C";
 
 Graph::~Graph() {
@@ -2301,7 +2301,7 @@ void Graph::new_axis() {
     }
     XYView* v = XYView::current_pick_view();
     erase_axis();
-    Coord x1, x2, y1, y2;
+    Coord x1{}, x2{}, y1{}, y2{};
     if (v) {
         v->zin(x1, y1, x2, y2);
     }
