@@ -17,7 +17,7 @@ class OcDeck: public OcGlyphContainer {
     virtual void save_action(const char*, Object*);
     virtual void flip_to(int);
     virtual void remove_last();
-    virtual void remove(int);
+    virtual void remove(long);
     virtual void move_last(int);  // make last item the i'th item
   private:
     OcDeckImpl* bi_;
