@@ -237,12 +237,7 @@ void parout() {
     if (brkpnt_exists) {
         brkpnt_str_ = "nrn_cur, nrn_jacob, nrn_state";
     } else {
-        brkpnt_str_ = "0, 0, 0";
-#if 1 || defined(__MINGW32__)
-        /* x86_64-w64-mingw32-gcc passed 0 without zeroing the high 32 bits */
-        /* also cygwin64 gcc 4.8.1, so cast to void* universally */
-        brkpnt_str_ = "(void*)0, (void*)0, (void*)0";
-#endif
+        brkpnt_str_ = "nullptr, nullptr, nullptr";
     }
 
     for (modbase = modprefix + strlen(modprefix); modbase != modprefix; modbase--) {
@@ -279,7 +274,7 @@ void parout() {
 \n#if METHOD3\nextern int _method3;\n#endif\n\
 \n#if !NRNGPU\
 \n#undef exp\
-\n#define exp hoc_Exp\nextern double hoc_Exp(double);\
+\n#define exp hoc_Exp\
 \n#endif\n\
 ");
     if (protect_include_) {
@@ -341,7 +336,7 @@ void parout() {
 	/*SUPPRESS 763*/\n\
 	/*SUPPRESS 765*/\n\
 	");
-    Lappendstr(defs_list, "extern double *getarg();\n");
+    Lappendstr(defs_list, "extern double *hoc_getarg(int);\n");
 #if VECTORIZE
     if (vectorize) {
         Sprintf(buf, "/* Thread safe. No static _p or _ppvar. */\n");
@@ -359,7 +354,6 @@ void parout() {
     Lappendstr(defs_list,
                "\n#if MAC\n#if !defined(v)\n#define v _mlhv\n#endif\n#if !defined(h)\n#define h "
                "_mlhh\n#endif\n#endif\n");
-    Lappendstr(defs_list, "\n#if defined(__cplusplus)\nextern \"C\" {\n#endif\n");
     Lappendstr(defs_list, "static int hoc_nrnpointerindex = ");
     if (pointercount) {
         q = nrnpointers->next;
@@ -429,11 +423,8 @@ extern Memb_func* memb_func;\n\
                    "\n"
                    "#define NMODL_TEXT 1\n"
                    "#if NMODL_TEXT\n"
-                   "static const char* nmodl_file_text;\n"
-                   "static const char* nmodl_filename;\n"
-                   "extern void hoc_reg_nmodl_text(int, const char*);\n"
-                   "extern void hoc_reg_nmodl_filename(int, const char*);\n"
-                   "#endif\n\n");
+                   "static void register_nmodl_text_and_filename(int mechtype);\n"
+                   "#endif\n");
     }
 
     /**** create special point process functions */
@@ -762,12 +753,12 @@ extern Memb_func* memb_func;\n\
     /*declaration of the range variables names to HOC */
     Lappendstr(
         defs_list,
-        "static void nrn_alloc(Prop*);\nstatic void  nrn_init(NrnThread*, _Memb_list*, int);\nstatic void nrn_state(NrnThread*, _Memb_list*, int);\n\
+        "static void nrn_alloc(Prop*);\nstatic void  nrn_init(NrnThread*, Memb_list*, int);\nstatic void nrn_state(NrnThread*, Memb_list*, int);\n\
 ");
     if (brkpnt_exists) {
         Lappendstr(defs_list,
-                   "static void nrn_cur(NrnThread*, _Memb_list*, int);\nstatic void  "
-                   "nrn_jacob(NrnThread*, _Memb_list*, int);\n");
+                   "static void nrn_cur(NrnThread*, Memb_list*, int);\nstatic void  "
+                   "nrn_jacob(NrnThread*, Memb_list*, int);\n");
     }
     /* count the number of pointers needed */
     ppvar_cnt = ioncount + diamdec + pointercount + areadec;
@@ -1138,7 +1129,7 @@ extern void hoc_register_tolerance(int, HocStateTolerance*, Symbol***);\n\
 extern void _cvode_abstol( Symbol**, double*, int);\n\n\
 ");
     Sprintf(buf,
-            "void _%s_reg() {\n\
+            "extern \"C\" void _%s_reg() {\n\
 	int _vectorized = %d;\n",
             modbase,
             vectorize);
@@ -1146,7 +1137,7 @@ extern void _cvode_abstol( Symbol**, double*, int);\n\n\
     q = lappendstr(defs_list, "");
     Lappendstr(defs_list, "_initlists();\n");
 #else
-    Sprintf(buf, "void _%s_reg() {\n	_initlists();\n", modbase);
+    Sprintf(buf, "extern \"C\" void _%s_reg() {\n	_initlists();\n", modbase);
     Lappendstr(defs_list, buf);
 #endif
 
@@ -1220,8 +1211,7 @@ extern void _cvode_abstol( Symbol**, double*, int);\n\n\
         }
         if (nmodl_text) {
             lappendstr(defs_list,
-                       "#if NMODL_TEXT\n  hoc_reg_nmodl_text(_mechtype, nmodl_file_text);\n  "
-                       "hoc_reg_nmodl_filename(_mechtype, nmodl_filename);\n#endif\n");
+                       "#if NMODL_TEXT\n  register_nmodl_text_and_filename(_mechtype);\n#endif\n");
         }
         sprintf(buf, " hoc_register_prop_size(_mechtype, %d, %d);\n", parraycount, ppvar_cnt);
         Lappendstr(defs_list, buf);
@@ -1447,7 +1437,7 @@ void ldifusreg() {
         dfdcur = STR(q);
         ++n;
         sprintf(buf,
-                "static void* _difspace%d;\nextern double nrn_nernst_coef();\n\
+                "static void* _difspace%d;\nextern double nrn_nernst_coef(int);\n\
 static double _difcoef%d(int _i, double* _p, Datum* _ppvar, double* _pdvol, double* _pdfcdc, Datum* _thread, NrnThread* _nt) {\n  \
  *_pdvol = ",
                 n,
@@ -2573,8 +2563,8 @@ static int _ode_count(int _type){ hoc_execerror(\"%s\", \"cannot be used with CV
                    "\n\
 static int _ode_count(int);\n\
 static void _ode_map(int, double**, double**, double*, Datum*, double*, int);\n\
-static void _ode_spec(NrnThread*, _Memb_list*, int);\n\
-static void _ode_matsol(NrnThread*, _Memb_list*, int);\n\
+static void _ode_spec(NrnThread*, Memb_list*, int);\n\
+static void _ode_matsol(NrnThread*, Memb_list*, int);\n\
 ");
         sprintf(buf,
                 "\n\
@@ -2588,7 +2578,7 @@ static int _ode_count(int _type){ return %d;}\n",
             cvode_proced_emit();
         } else {
             Lappendstr(procfunc,
-                       "\nstatic void _ode_spec(NrnThread* _nt, _Memb_list* _ml, int _type) {\n");
+                       "\nstatic void _ode_spec(NrnThread* _nt, Memb_list* _ml, int _type) {\n");
             out_nt_ml_frag(procfunc);
             lst = get_ion_variables(1);
             if (lst->next->itemtype)
@@ -2670,7 +2660,7 @@ static void _ode_synonym(int _cnt, double** _pp, Datum** _ppd) {");
             }
             Lappendstr(procfunc, "}\n");
             Lappendstr(procfunc,
-                       "\nstatic void _ode_matsol(NrnThread* _nt, _Memb_list* _ml, int _type) {\n");
+                       "\nstatic void _ode_matsol(NrnThread* _nt, Memb_list* _ml, int _type) {\n");
             out_nt_ml_frag(procfunc);
             lst = get_ion_variables(1);
             if (lst->next->itemtype)
@@ -2814,9 +2804,8 @@ void net_receive(Item* qarg, Item* qp1, Item* qp2, Item* qstmt, Item* qend) {
             insertstr(qstmt, " assert(_tsav <= t); _tsav = t;");
         } else {
             insertstr(qstmt,
-                      " if (_tsav > t){ extern char* hoc_object_name(); "
-                      "hoc_execerror(hoc_object_name(_pnt->ob), \":Event arrived out of order. "
-                      "Must call ParallelContext.set_maxstep AFTER assigning minimum "
+                      " if (_tsav > t){ hoc_execerror(hoc_object_name(_pnt->ob), \":Event arrived "
+                      "out of order. Must call ParallelContext.set_maxstep AFTER assigning minimum "
                       "NetCon.delay\");}\n _tsav = t;");
         }
     }
