@@ -19,9 +19,9 @@
 #include <hoclist.h>
 #include <ocpointer.h>
 #include <ocptrvector.h>
+#include "treeset.h"
 
 void nrniv_recalc_ptrs();
-extern "C" double* nrn_recalc_ptr(double*);
 extern NetCvode* net_cvode_instance;
 
 extern void nrn_linmod_update_ptrs(void*);
