@@ -17,6 +17,19 @@ typedef void (*pnt_receive_init_t)(Point_process*, double*, double);
 extern Prop* need_memb_cl(Symbol*, int*, int*);
 extern Prop* prop_alloc(Prop**, int, Node*);
 
+void artcell_net_send(void** v, double* weight, Point_process* pnt, double td, double flag);
+[[deprecated("non-void* overloads are preferred")]] void artcell_net_send(void** v,
+                                                                          double* weight,
+                                                                          void* pnt,
+                                                                          double td,
+                                                                          double flag);
+void nrn_net_send(void** v, double* weight, Point_process* pnt, double td, double flag);
+[[deprecated("non-void* overloads are preferred")]] void nrn_net_send(void** v,
+                                                                      double* weight,
+                                                                      void* pnt,
+                                                                      double td,
+                                                                      double flag);
+
 extern double nrn_ion_charge(Symbol*);
 #if defined(__cplusplus)
 extern "C" {
@@ -126,7 +139,6 @@ extern Prop* nrn_point_prop_;
 
 extern int at_time(NrnThread*, double);
 
-extern void artcell_net_send(void**, double*, Point_process*, double, double);
 extern void artcell_net_move(void**, Point_process*, double);
 
 extern int ifarg(int);
@@ -140,7 +152,6 @@ extern int v_structure_change;
 extern pnt_receive_t* pnt_receive;
 extern pnt_receive_init_t* pnt_receive_init;
 extern short* pnt_receive_size;
-extern void nrn_net_send(void**, double*, Point_process*, double, double);
 extern void nrn_net_event(Point_process*, double);
 extern void nrn_net_move(void**, Point_process*, double);
 
