@@ -1,27 +1,12 @@
-#ifndef mcran4_h
-#define mcran4_h
+#pragma once
+#include <cstdint>
 
-#include <nrnconf.h>
-#if defined(HAVE_STDINT_H)
-#include <stdint.h>
-#endif
-
-#if defined(__cplusplus)
-extern "C" {
-#endif
-
-extern void mcell_ran4_init(uint32_t);
-extern double mcell_ran4(uint32_t* high, double* x, unsigned int n, double range);
-extern double mcell_ran4a(uint32_t* idx1);
-extern uint32_t mcell_iran4(uint32_t* idx1);
-extern double nrnRan4dbl(uint32_t* idx1, uint32_t idx2);
-extern uint32_t nrnRan4int(uint32_t* idx1, uint32_t idx2);
-
-#if defined(__cplusplus)
-}
-#endif
-
-#endif
+void mcell_ran4_init(uint32_t);
+double mcell_ran4(uint32_t* idx1, double* x, unsigned int n, double range);
+double mcell_ran4a(uint32_t* idx1);
+uint32_t mcell_iran4(uint32_t* idx1);
+double nrnRan4dbl(uint32_t* idx1, uint32_t idx2);
+uint32_t nrnRan4int(uint32_t* idx1, uint32_t idx2);
 
 /*
  The original ran4 generator was copyrighted by "Numerical Recipes in C"
