@@ -59,7 +59,11 @@ for line in text.splitlines():
     names = line.replace(",", " ").replace(";", " ").split()
     if not skip and len(names) > 2:
         print(line)  # entire filtered neuron.h file without pgcc added lines.
-        process(names[1], names[2:])
+        assert names[0] == "extern"
+        names.pop(0)
+        if names[0] in {'"C"', '"C++"'}:
+            names.pop(0)
+        process(names[0], names[1:])
 
 
 print(
